Added new ui shader
